pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "fpvs-studio"
version = "0.1.0"
description = "FPVS Studio scaffolding and engine-neutral domain layer."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10,<3.13"
authors = [
  { name = "FPVS Studio contributors" },
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Scientific/Engineering",
]
dependencies = [
  "numpy>=1.26,<3.0",
  "pillow>=10,<12",
  "pydantic>=2.8,<3.0",
]

[project.optional-dependencies]
gui = [
  "PySide6>=6.7,<7.0",
]
engine = [
  "psychopy>=2024.2,<2027.0",
]
dev = [
  "mypy>=1.10,<2.0",
  "pytest>=8.2,<9.0",
  "ruff>=0.6,<1.0",
]

[project.scripts]
fpvs-studio = "fpvs_studio.app.main:main"

[tool.pytest.ini_options]
addopts = "-ra"
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "B",
  "UP",
]

[tool.mypy]
python_version = "3.10"
mypy_path = "src"
pretty = true
show_error_codes = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
no_implicit_optional = true
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_defs = true

[tool.hatch.build.targets.wheel]
packages = ["src/fpvs_studio"]
